requirements.txt
aiogram==2.14.3
python-dotenv==1.0.0
SQLAlchemy==1.4.41
psycopg2-binary==2.9.6
fastapi==0.104.1
uvicorn==0.24.0
requests==2.28.2
jinja2==3.1.2
python-multipart==0.0.6
aiofiles==23.2.1
matplotlib==3.7.1
aioschedule==0.5.2 